// shared/types/addressTypes.d.ts
export interface RegionInfo {
    regionID?:string
    regionName: string
    regionShort: string

}

export interface SubcityInfo {
    subcityID?:string
    subcityName: string
    subcityShort: string
    
}

export interface WoredaInfo {
    woredaID?: string
    woredaName: string
    woredaShort?:string
}